requirements.txt
PyYAML>=6.0
pynput>=1.7.6
